pyproject.toml
[project]
name = "hotel-price-scraper-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyside6>=6.9.1",
    "selenium>=4.35.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
    "undetected-chromedriver>=3.5.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
